c++: more porting

+++ lib/hooks/restart.c
@@ -54,7 +54,7 @@ static int restart_stop(struct hook *h)
 
 static int restart_read(struct hook *h, struct sample *smps[], unsigned *cnt)
 {
-	int i;
+	unsigned i;
 	struct restart *r = (struct restart *) h->_vd;
 	struct sample *prev, *cur = NULL;
 
@@ -65,7 +65,7 @@ static int restart_read(struct hook *h, struct sample *smps[], unsigned *cnt)
 
 		if (prev) {
 			/* A wrap around of the sequence no should not be treated as a simulation restart */
-			if (cur->sequence == 0 && prev->sequence <= UINT32_MAX - 128) {
+			if (cur->sequence == 0 && prev->sequence <= (int) (UINT32_MAX - 128)) {
 				warn("Simulation from node %s restarted (previous->seq=%u, current->seq=%u)",
 					node_name(h->node), prev->sequence, cur->sequence);
